CMake: build boost unit tests of tuttleHost
* Add CMakeList.txt in libraries/tuttle/tests, which creates a test for
each source file in directory.
* Need to rename some test files in order to have different name (not
main.cpp for all tests).
* All source files in tests directory need to be a test (CMake will try
to create an executable for each one of them).
* The build generates a new folder (libraries/tuttle/tests/testBin) with
all binaries to test tuttleHost.
