--- libraries/tuttle/tests/asynchronous/SConscript
@@ -1,16 +1,0 @@
-Import( 'project', 'libs' )
-
-project.UnitTest(
-	target = project.getDirs([-3,-1]),
-	dirs = ['.'],
-	libraries = [
-		libs.tuttleTest,
-		],
-	execLibraries = [
-		libs.pluginJpeg,
-		libs.pluginPng,
-		libs.pluginExr,
-		libs.pluginInvert,
-		],
-	)
-
