--- libraries/tuttle/tests/asynchronous/main.cpp
@@ -1,2 +1,0 @@
-#define BOOST_TEST_MODULE tuttle_asynchronous
-#include <tuttle/test/main.hpp>
