--- libraries/tuttle/tests/graph/SConscript
@@ -1,19 +1,0 @@
-Import( 'project', 'libs' )
-
-project.UnitTest(
-	target = project.getDirs([-3,-1]),
-	dirs = ['.'],
-	libraries = [
-		libs.tuttleTest,
-		],
-	execLibraries = [
-		libs.pluginJpeg,
-		libs.pluginPng,
-		libs.pluginExr,
-		libs.pluginInvert,
-		libs.pluginMerge,
-		libs.pluginCrop,
-		libs.pluginBlur,
-		],
-	)
-
