--- libraries/tuttle/tests/graph/main.cpp
@@ -1,2 +1,0 @@
-#define BOOST_TEST_MODULE tuttle_graph
-#include <tuttle/test/main.hpp>
