--- libraries/tuttle/tests/internalGraph/dummy/DummyAttribute.cpp
@@ -1,8 +1,0 @@
-#include "DummyAttribute.hpp"
-
-std::ostream& operator<<( std::ostream& os, const DummyAttribute& v )
-{
-	os << v.name() ;
-	return os;
-}
-
