--- libraries/tuttle/tests/internalGraph/dummy/DummyEdge.cpp
@@ -1,13 +1,0 @@
-#include "DummyEdge.hpp"
-
-namespace tuttle {
-namespace test {
-
-std::ostream& operator<<( std::ostream& os, const DummyEdge& v )
-{
-	os << v.getName() ;
-	return os;
-}
-
-} // namespace test
-} // namespace tuttle
