--- libraries/tuttle/tests/memory/SConscript
@@ -1,10 +1,0 @@
-Import( 'project', 'libs' )
-
-project.UnitTest(
-	target=project.getDirs([-3,-1]),
-	dirs=['.'],
-	libraries = [
-		libs.tuttleTest,
-		]
-	)
-
