+++ libraries/tuttle/tests/memory/memory.cpp
@@ -1,11 +1,11 @@
+#define BOOST_TEST_MODULE tuttle_memory
+#include <tuttle/test/main.hpp>
+
 // custom host
 #include <tuttle/host/memory/MemoryPool.hpp>
 #include <tuttle/host/memory/MemoryCache.hpp>
 
 #include <iostream>
-
-#define BOOST_TEST_MODULE tuttle_memory
-#include <tuttle/test/unit_test.hpp>
 
 using namespace boost::unit_test;
 using namespace std;
@@ -16,15 +16,13 @@ BOOST_AUTO_TEST_SUITE( memory_tests_suite01 )
 BOOST_AUTO_TEST_CASE( memoryPool )
 {
 	{
-		memory::MemoryPool emptyPool(0);
+		// empty pool
+		memory::MemoryPool emptyPool;
 		BOOST_CHECK_EQUAL( 0U, emptyPool.getMaxMemorySize() );
 		BOOST_CHECK_EQUAL( 0U, emptyPool.getUsedMemorySize() );
 		BOOST_CHECK_EQUAL( 0U, emptyPool.getAllocatedMemorySize() );
 		BOOST_CHECK_EQUAL( 0U, emptyPool.getWastedMemorySize() );
-		BOOST_REQUIRE_THROW(
-		{
-			 emptyPool.allocate( 10 );
-		}, std::exception );
+		BOOST_REQUIRE_THROW( emptyPool.allocate( 10 ), std::exception );
 	}
 	
 	memory::MemoryPool pool( 30 );
@@ -33,6 +31,9 @@ BOOST_AUTO_TEST_CASE( memoryPool )
 	BOOST_CHECK_EQUAL( 0U, pool.getUsedMemorySize() );
 	BOOST_CHECK_EQUAL( 0U, pool.getAllocatedMemorySize() );
 	BOOST_CHECK_EQUAL( 0U, pool.getWastedMemorySize() );
+
+	// can't allocate more than size of pool
+	BOOST_REQUIRE_THROW( pool.allocate( 50 ), std::exception );
 
 	{
 		// first allocation
@@ -54,19 +55,30 @@ BOOST_AUTO_TEST_CASE( memoryPool )
 		BOOST_CHECK_EQUAL( 0U, pool.getWastedMemorySize() );
 	}
 	{
-		// same thing for a smaller allocation
-		const memory::IPoolDataPtr pData = pool.allocate( 1 );
-		BOOST_CHECK_EQUAL( 1U, pData->size() );
+		// same thing for a smaller allocation (but bigger than half of the allocated buffer)
+		// resuse memory
+		const memory::IPoolDataPtr pData = pool.allocate( 6 );
+		BOOST_CHECK_EQUAL( 6U, pData->size() );
 		BOOST_CHECK_EQUAL( 10U, pData->reservedSize() );
 		BOOST_CHECK_EQUAL( 10U, pool.getUsedMemorySize() );
 		BOOST_CHECK_EQUAL( 10U, pool.getAllocatedMemorySize() );
-		BOOST_CHECK_EQUAL( 9U, pool.getWastedMemorySize() );
+		BOOST_CHECK_EQUAL( 4U, pool.getWastedMemorySize() );
+	}
+	{
+		// same thing for a smaller allocation (smaller than half of the allocated buffer)
+		// allocate new buffer
+		const memory::IPoolDataPtr pData = pool.allocate( 4 );
+		BOOST_CHECK_EQUAL( 4U, pData->size() );
+		BOOST_CHECK_EQUAL( 4U, pData->reservedSize() );
+		BOOST_CHECK_EQUAL( 4U, pool.getUsedMemorySize() );
+		BOOST_CHECK_EQUAL( 14U, pool.getAllocatedMemorySize() );
+		BOOST_CHECK_EQUAL( 0U, pool.getWastedMemorySize() );
 	}
 	{
 		// asking for a bigger alloc will alloc this time
 		const memory::IPoolDataPtr pData = pool.allocate( 15 );
 		BOOST_CHECK_EQUAL( 15U, pool.getUsedMemorySize() );
-		BOOST_CHECK_EQUAL( 25U, pool.getAllocatedMemorySize() );
+		BOOST_CHECK_EQUAL( 29U, pool.getAllocatedMemorySize() );
 		BOOST_CHECK_EQUAL( 0U, pool.getWastedMemorySize() );
 	}
 	{
@@ -79,18 +91,24 @@ BOOST_AUTO_TEST_CASE( memoryPool )
 
 		// no new allocation should have take place
 		BOOST_CHECK_EQUAL( 10U, pool.getUsedMemorySize() );
-		BOOST_CHECK_EQUAL( 25U, pool.getAllocatedMemorySize() );
+		BOOST_CHECK_EQUAL( 29U, pool.getAllocatedMemorySize() );
 
 		const char* const_data = pData->data();
 		char* data             = pData->data();
 		data[0] = const_data[0];
 	}
-	BOOST_REQUIRE_THROW( pool.allocate( 50 ), std::exception );
+	{
+		// clear memory allocated
+		BOOST_CHECK_EQUAL( 0U, pool.getUsedMemorySize() );
+		BOOST_CHECK_EQUAL( 29U, pool.getAllocatedMemorySize() );
+		pool.clear();
+		BOOST_CHECK_EQUAL( 0U, pool.getUsedMemorySize() );
+		BOOST_CHECK_EQUAL( 0U, pool.getAllocatedMemorySize() );
+	}
 }
 
 BOOST_AUTO_TEST_CASE( memoryCache )
 {
-	memory::MemoryPool pool;
 	memory::MemoryCache cache;
 
 	// checking initial state
@@ -134,4 +152,3 @@ BOOST_AUTO_TEST_CASE( memoryCache )
 }
 
 BOOST_AUTO_TEST_SUITE_END()
-
