--- libraries/tuttle/tests/ofx/clone/SConscript
@@ -1,11 +1,0 @@
-Import( 'project', 'libs' )
-
-project.UnitTest(
-	target = project.getDirs([-4,-2,-1]),
-	dirs = ['.'],
-	libraries = [
-		libs.tuttleHost,
-		libs.boost_unit_test_framework,
-		]
-	)
-
