+++ libraries/tuttle/tests/ofx/properties/properties.cpp
@@ -1,5 +1,5 @@
-#define BOOST_TEST_MODULE properties_tests
-#include <tuttle/test/unit_test.hpp>
+#define BOOST_TEST_MODULE properties_properties_tests
+#include <tuttle/test/main.hpp>
 
 #include <tuttle/common/utils/global.hpp>
 #include <tuttle/host/Core.hpp>
