+++ libraries/tuttle/tests/ofx/properties/serialization.cpp
@@ -1,5 +1,5 @@
-//#define BOOST_TEST_MODULE properties_tests
-#include <tuttle/test/unit_test.hpp>
+#define BOOST_TEST_MODULE properties_sezialization_tests
+#include <tuttle/test/main.hpp>
 
 #include <tuttle/common/utils/global.hpp>
 #include <tuttle/host/Core.hpp>
