--- libraries/tuttle/tests/time/SConscript
@@ -1,18 +1,0 @@
-Import( 'project', 'libs' )
-
-project.UnitTest(
-	target = project.getDirs([-3,-1]),
-	dirs = ['.'],
-	libraries = [
-		libs.tuttleHost,
-		libs.boost_unit_test_framework,
-		],
-	execLibraries = [
-		libs.pluginJpeg,
-		libs.pluginPng,
-		libs.pluginExr,
-		libs.pluginInvert,
-		libs.pluginTimeShift,
-		],
-	)
-
